Create CALCSRCHSH v00.01
This program will generate an MD5 hash on the contents of an IBM i source member. 

You may note that the program trims blanks after the last character in SRCDTA.  The reason I did this is because source files may be defined with different lengths, and the extra trailing blanks in the longer source member will generate a different hash than a shorter source member, even if the two source members are otherwise identical.

If SRCDTA is blank, I substitute a hex'FF' so that a different number of blank lines in the source member will generate a different hash.
This is a working prototype of what will eventually become a service program.  

This is my first project on GitHub, so I don't know how this will turn out.